Improved and consolidated image Jinja tag logic
This commit consolidates and relocates our Jinja `image_alt_text` tag
into a new v1.jinja2tags.images module. It also redefines the default
Wagtail `image` tag so that it can optionally take an image dict as
its argument instead of always needing to take an Image object.

The Wagtail {% image %} tag [0] takes an image object and generates an
<img> tag in HTML. We use this in various Jinja templates for modules
that contain images. And it's very useful when you have an image object:

{% set img = image(page.image, "original") %}
<img src="{{ img.url }}">

But sometimes we may want to call one of our templates and render an
image but don't actually have an Image object, for example if we want
to hardcode the image link elements (as we do currently on the home page
and on the TCCP landing page). The above code doesn't work well because
the {% image %} tag can't handle something that isn't actually an Image.

This commit thus replaces Wagtail's {% image %} with our own {% image %}
that can handle that case; it acts as a passthrough or no-op that simply
returns the object that was passed in. That lets you do this:

image_dict = {"url": "cat.png"}

{% set img = image(image_dict, "original") %}
<img src="{{ img.url }}">

https://docs.wagtail.org/en/stable/reference/jinja2.html#image
