+++ cfgov/tccp/jinja2/tccp/landing_page.html
@@ -33,8 +33,7 @@
             "width": 1140
         },
         "background_color": "#FFE1B9"
-    },
-    image = image_passthrough
+    }
 %}
     {% include "v1/includes/molecules/hero.html" with context %}
 {% endwith %}
