+++ cfgov/v1/jinja2tags/__init__.py
@@ -2,39 +2,13 @@ from jinja2 import pass_context
 from jinja2.ext import Extension
 
 from core.feature_flags import environment_is
-from v1.jinja2tags.datetimes import DatetimesExtension
-from v1.jinja2tags.fragment_cache import FragmentCacheExtension
-from v1.models.images import CFGOVRendition
 from v1.templatetags.app_urls import app_page_url, app_url
 from v1.util import ref
 from v1.util.util import get_unique_id
 
-
-def image_alt_value(image):
-    """Given an ImageBasic block or a CFGOVImage rendition as `image`,
-    return the appropriate alt text.
-
-    Return the CFGOVImage rendition's alt field, if present.
-    Returns the alt field from the block, if it is set.
-    Otherwise, returns the alt field from the CFGOVImage object, if it is set.
-    Otherwise, returns an empty string.
-    """
-
-    # Check to see if the passed value is a CFGOVRendition
-    if isinstance(image, CFGOVRendition):
-        return image.alt
-
-    # Otherwise, if it is a block
-    if image and hasattr(image, "get"):
-        block_alt = image.get("alt")
-        upload = image.get("upload")
-
-        if block_alt:
-            return block_alt
-        elif upload and upload.alt:
-            return upload.alt
-
-    return ""
+from .datetimes import DatetimesExtension
+from .fragment_cache import FragmentCacheExtension
+from .images import ImagesExtension
 
 
 def is_filter_selected(context, fieldname, value):
@@ -115,7 +89,6 @@ class V1Extension(Extension):
                 "choices_for_page_type": ref.choices_for_page_type,
                 "get_category_icon": ref.get_category_icon,
                 "get_unique_id": get_unique_id,
-                "image_alt_value": image_alt_value,
                 "is_blog": ref.is_blog,
                 "is_report": ref.is_report,
                 "is_filter_selected": pass_context(is_filter_selected),
@@ -130,4 +103,5 @@ class V1Extension(Extension):
 # Nicer import names
 datetimes_extension = DatetimesExtension
 fragment_cache_extension = FragmentCacheExtension
+images_extension = ImagesExtension
 v1_extension = V1Extension
